Modified identical answer finder
The old identical answer finder is not good enough because chatbot change some words slightly so the it not 100% identical strings but it's 100% identical answer to question.
Ex:-  Me: What is the speed of the light
        Bot: What's the speed of the light?
Those 2 are not identical strings but It says the same thing. So I changed is_answer_identical function to first separate words and then check how many similar words are in the 2 sentences and assign a weighted penalty by the number of similar words. I tested this function with my chat log with the Charles on Twitter. It work fine but I don't know how well this work on production server.
One more thing I think it's good if Charles stop giving any link. most of them are broken links and others are irrelevant.